+++ app/lib/scorecard/models.dart
@@ -330,7 +330,13 @@ class PanaReport implements ReportData {
   /// List of tags computed by `pana`.
   final List<String> derivedTags;
 
+  /// WARNING: this field is deprecated, use `allDependencies` instead.
+  /// TODO: remove this field after the runtime `2020.11.27` gets released.
   final List<PkgDependency> pkgDependencies;
+
+  /// The list of packages that the current one depends on either directly or
+  /// transitively.
+  final List<String> allDependencies;
 
   final LicenseFile licenseFile;
 
@@ -348,6 +354,7 @@ class PanaReport implements ReportData {
     @required this.reportStatus,
     @required this.derivedTags,
     @required this.pkgDependencies,
+    @required this.allDependencies,
     @required this.licenseFile,
     @required this.report,
     @required this.flags,
@@ -358,6 +365,10 @@ class PanaReport implements ReportData {
 
   @override
   Map<String, dynamic> toJson() => _$PanaReportToJson(this);
+
+  // TODO: remove when pkgDependencies is removed
+  List<String> get allDependenciesWithFallback =>
+      allDependencies ?? pkgDependencies?.map((p) => p.package)?.toList();
 }
 
 @JsonSerializable()