+++ pkg/fake_pub_server/lib/fake_pub_server.dart
@@ -16,9 +16,7 @@ import 'package:shelf/shelf_io.dart';
 import 'package:pub_dev/account/backend.dart';
 import 'package:pub_dev/account/testing/fake_auth_provider.dart';
 import 'package:pub_dev/frontend/handlers.dart';
-import 'package:pub_dev/frontend/testing/fake_upload_signer_service.dart';
 import 'package:pub_dev/package/name_tracker.dart';
-import 'package:pub_dev/package/upload_signer_service.dart';
 import 'package:pub_dev/publisher/domain_verifier.dart';
 import 'package:pub_dev/publisher/testing/fake_domain_verifier.dart';
 import 'package:pub_dev/service/services.dart';
@@ -46,8 +44,6 @@ class FakePubServer {
         fn: () async {
           registerAuthProvider(FakeAuthProvider(port));
           registerDomainVerifier(FakeDomainVerifier());
-          registerUploadSigner(
-              FakeUploadSignerService(configuration.storageBaseUrl));
 
           nameTracker.startTracking();
           spamBackend.setSpamConfig(spamWords: ['SPAM-SPAM-SPAM']);
@@ -58,13 +54,7 @@ class FakePubServer {
           final server = await IOServer.bind('localhost', port);
           serveRequests(server.server, (request) async {
             return await ss.fork(() async {
-              final rs = await extraHandler(request) ?? await handler(request);
-              try {
-                return await _validateAndWrapResponse(request, rs) ?? rs;
-              } catch (e, st) {
-                _logger.severe('Response validation failed', e, st);
-                return shelf.Response.internalServerError();
-              }
+              return await extraHandler(request) ?? await handler(request);
             }) as shelf.Response;
           });
           _logger.info('running on port $port');
@@ -77,18 +67,4 @@ class FakePubServer {
         });
     _logger.info('closed');
   }
-}
-
-// Normally this method could be just void, however reading the shelf.Response's
-// body forces us to return a new instance of the response.
-Future<shelf.Response> _validateAndWrapResponse(
-    shelf.Request rq, shelf.Response rs) async {
-  if (rs == null) return null;
-
-  final ct = rs.headers[HttpHeaders.contentTypeHeader];
-  if (ct == null || ct.isEmpty) {
-    // TODO: force headers everywhere and throw exception instead of logging
-    _logger.warning('Content type header is missing for ${rq.requestedUri}.');
-  }
-  return null;
 }