+++ CHANGELOG.md
@@ -2,6 +2,11 @@ Important changes to data-models, configuration and migrations between each
 AppEngine version, listed here to ease deployment and troubleshooting.
 
 ## Next Release (replace with git tag when deployed)
+ * Bumped runtimeVersion to `2020.11.27`.
+ * NOTE: `PanaReport` has deprecated `pkgDependencies`.
+         The field can be removed after this release is stable.
+ * NOTE: added `Package`'s `latestPublished` and `latestPrereleasePublished`.
+         TODO(deferred): schedule backfill after this release.
 
 ## `20201126t141046-all`
  * Bumped runtimeVersion to `2020.11.25`.